requirements.txt
rich
python-whois
shodan
requests
pyhunter
beautifulsoup4
lxml
googlesearch-python
PyPDF2
Pillow
pynput
dnspython
paramiko
pysmb
pymongo
redis
opencv-python
pyperclip
psutil
sounddevice
scipy
browser-history
scapy
sounddevice
scipy
